modify swagger file finder to use new dir structure
