+++ config/application.rb
@@ -23,7 +23,8 @@ module EventCaptureApi
     config.load_defaults 5.2
 
     config.eager_load_paths << Rails.root.join('lib')
-    config.eager_load_paths << Rails.root.join("datum", "schema", "org", "openstax")
+    config.eager_load_paths << Rails.root.join("schemas", "org", "openstax")
+    config.autoload_paths << Rails.root.join("schemas", "org", "openstax")
 
     # Settings in config/environments/* take precedence over those specified here.
     # Application configuration can go into files in config/initializers
