+++ config/secrets.yml.example
@@ -14,7 +14,7 @@ development:
     delivery_threshold: 5
     delivery_interval: 5
     schema_url: 'http://localhost:8081/'
-    schemas_path: 'app/schemas'
+    schemas_path: 'schemas'
 test:
   accounts:
     sso:
@@ -31,7 +31,7 @@ test:
     delivery_threshold: 5
     delivery_interval: 5
     schema_url: 'http://localhost:8081/'
-    schemas_path: 'app/schemas'
+    schemas_path: 'schemas'
 production:
   accounts:
     sso:
