+++ docker/docker-compose.yml
@@ -2,12 +2,14 @@
 version: "2"
 services:
   api:
-    build: 
+    build:
       context: ..
       dockerfile: docker/Dockerfile
     ports:
       - "3001:3001"
-    command: bash -c "rake about && bin/rails server -b '0.0.0.0' -p 3001"
+    command: bash -c "rm -f tmp/pids/server.pid && bin/rake about && bin/rails server -b '0.0.0.0' -p 3001"
+    stdin_open: true  # allow for debugger
+    tty: true         # interaction
     volumes:
       - ..:/code
       - /code/tmp
