+++ spec/lib/swagger_finder_spec.rb
@@ -1,0 +1,25 @@
+# frozen_string_literal: true
+
+require 'rails_helper'
+
+RSpec.describe SwaggerFinder do
+  before do
+    allow(Rails.application.secrets).to receive(:kafka).and_return({schemas_path: 'schemas'})
+  end
+
+  describe '#files' do
+    subject(:files) { described_class.new.files }
+
+    it 'finds the files' do
+      expect(files.first.include?('swagger')).to be_truthy
+    end
+  end
+
+  describe '#classes' do
+    subject(:classes) { described_class.new.classes }
+
+    it 'finds the classes' do
+      expect(classes.first.to_s.include?('Nudged::V1::Swagger')).to be_truthy
+    end
+  end
+end
